requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
loguru==0.7.3
openai==2.15.0
prettytable==3.17.0
pydantic==2.12.5
pydantic_core==2.41.5
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.45
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
win32_setctime==1.2.0
